Progress on AddURL form
Pressing OK validates the entered URL. If it is valid, a save dialog
pops up. If it isn't, an error appears. If a filename is entered (ie.
save dialog isn't cancelled) then an entry is added into the table on
Form1. If it isn't, then it just goes back to how it was before pressing
ok. No actual saving or anything yet of course.